+++ sketch-core/src/androidInstrumentedTest/kotlin/com/github/panpf/sketch/core/android/test/util/PlatformContextsAndroidTest.kt
@@ -4,7 +4,7 @@ import android.app.ActivityManager
 import android.content.Context
 import android.content.pm.ApplicationInfo
 import com.github.panpf.sketch.test.utils.getTestContext
-import com.github.panpf.sketch.util.totalAvailableMemoryBytes
+import com.github.panpf.sketch.util.maxMemory
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
@@ -16,17 +16,20 @@ class PlatformContextsAndroidTest {
     }
 
     @Test
-    fun testTotalAvailableMemoryBytes() {
+    fun testMaxMemory() {
         val context = getTestContext()
-        val activityManager = context.getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager?
-        val isLargeHeap =
-            (context.applicationInfo.flags and ApplicationInfo.FLAG_LARGE_HEAP) != 0
-        val appMemoryClassBytes = when {
-            activityManager != null && isLargeHeap -> activityManager.largeMemoryClass * 1024L * 1024L
-            activityManager != null && !isLargeHeap -> activityManager.memoryClass * 1024L * 1024L
-            else -> 16 * 1024L * 1024L
+        val memoryClassMegabytes = try {
+            val activityManager =
+                context.getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager
+            val isLargeHeap =
+                (context.applicationInfo.flags and ApplicationInfo.FLAG_LARGE_HEAP) != 0
+            if (isLargeHeap) activityManager.largeMemoryClass else activityManager.memoryClass
+        } catch (e: Throwable) {
+            e.printStackTrace()
+            128
         }
-        assertEquals(appMemoryClassBytes, context.totalAvailableMemoryBytes())
+        val appMemoryClassBytes = memoryClassMegabytes * 1024L * 1024L
+        assertEquals(appMemoryClassBytes, context.maxMemory())
     }
 
     @Test