+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/util/SystemCallbacks.common.kt
@@ -22,7 +22,9 @@ import com.github.panpf.sketch.Sketch
  * Create a platform-specific [SystemCallbacks] instance
  *
  * @see com.github.panpf.sketch.core.android.test.util.SystemCallbacksAndroidTest.testSystemCallbacks
- * @see com.github.panpf.sketch.core.nonandroid.test.util.SystemCallbacksNonAndroidTest.testSystemCallbacks
+ * @see com.github.panpf.sketch.core.desktop.test.util.SystemCallbacksDesktopTest.testSystemCallbacks
+ * @see com.github.panpf.sketch.core.jscommon.test.util.SystemCallbacksJsCommonTest.testSystemCallbacks
+ * @see com.github.panpf.sketch.core.ios.test.util.SystemCallbacksIosTest.testSystemCallbacks
  */
 internal expect fun SystemCallbacks(sketch: Sketch): SystemCallbacks
 
@@ -30,7 +32,9 @@ internal expect fun SystemCallbacks(sketch: Sketch): SystemCallbacks
  * Monitor network connection and system status
  *
  * @see com.github.panpf.sketch.core.android.test.util.SystemCallbacksAndroidTest.testAndroidSystemCallbacks
- * @see com.github.panpf.sketch.core.nonandroid.test.util.SystemCallbacksNonAndroidTest.testNoopSystemCallbacks
+ * @see com.github.panpf.sketch.core.desktop.test.util.SystemCallbacksDesktopTest.testDesktopSystemCallbacks
+ * @see com.github.panpf.sketch.core.jscommon.test.util.SystemCallbacksJsCommonTest.testJsSystemCallbacks
+ * @see com.github.panpf.sketch.core.ios.test.util.SystemCallbacksIosTest.testIosSystemCallbacks
  */
 interface SystemCallbacks {