+++ sketch-core/src/desktopMain/kotlin/com/github/panpf/sketch/util/SystemCallbacks.desktop.kt
@@ -1,0 +1,77 @@
+/*
+ * Copyright (C) 2024 panpf <panpfpanpf@outlook.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.github.panpf.sketch.util
+
+import com.github.panpf.sketch.Sketch
+import kotlinx.atomicfu.atomic
+import java.lang.management.ManagementFactory
+import java.lang.management.MemoryNotificationInfo
+import java.lang.management.MemoryType
+import javax.management.NotificationEmitter
+import javax.management.NotificationListener
+import kotlin.math.roundToLong
+
+/**
+ * Create an instance of [SystemCallbacks] for desktop platforms
+ *
+ * @see com.github.panpf.sketch.core.desktop.test.util.SystemCallbacksDesktopTest.testSystemCallbacks
+ */
+internal actual fun SystemCallbacks(sketch: Sketch): SystemCallbacks =
+    DesktopSystemCallbacks(sketch)
+
+/**
+ * Noop implementation of [SystemCallbacks]
+ *
+ * @see com.github.panpf.sketch.core.desktop.test.util.SystemCallbacksDesktopTest.testDesktopSystemCallbacks
+ */
+private class DesktopSystemCallbacks(val sketch: Sketch) : SystemCallbacks {
+
+    // TODO Implement network type detection for desktop platforms.
+    //  https://github.com/jordond/connectivity/blob/main/connectivity-apple/src/appleMain/kotlin/dev/jordond/connectivity/internal/AppleConnectivityProvider.kt
+    override val isCellularNetworkConnected get() = false
+
+    private val _isShutdown = atomic(false)
+    override var isShutdown: Boolean by _isShutdown
+
+    private val listener = NotificationListener { notification, _ ->
+        if (notification.type == MemoryNotificationInfo.MEMORY_THRESHOLD_EXCEEDED) {
+            sketch.memoryCache.clear()
+        }
+    }
+    private val oldGenPool by lazy {
+        ManagementFactory.getMemoryPoolMXBeans()
+            .find { it.type == MemoryType.HEAP && it.name.contains("Old") }
+    }
+
+    override fun register() {
+        val oldGenPool = oldGenPool
+        if (oldGenPool != null) {
+            oldGenPool.usageThreshold = (oldGenPool.usage.max * 0.8).roundToLong()
+            val emitter = oldGenPool as NotificationEmitter
+            emitter.addNotificationListener(listener, null, null)
+        }
+    }
+
+    override fun shutdown() {
+        if (_isShutdown.getAndSet(true)) return
+        val oldGenPool = oldGenPool
+        if (oldGenPool != null) {
+            val emitter = oldGenPool as NotificationEmitter
+            emitter.removeNotificationListener(listener)
+        }
+    }
+}