+++ sketch-core/src/desktopTest/kotlin/com/github/panpf/sketch/core/desktop/test/util/PlatformContextsDesktopTest.kt
@@ -1,8 +1,19 @@
 package com.github.panpf.sketch.core.desktop.test.util
 
+import com.github.panpf.sketch.PlatformContext
+import com.github.panpf.sketch.util.maxMemory
 import kotlin.test.Test
+import kotlin.test.assertEquals
 
 class PlatformContextsDesktopTest {
+
+    @Test
+    fun testMaxMemory() {
+        assertEquals(
+            expected = Runtime.getRuntime().maxMemory(),
+            actual = PlatformContext.INSTANCE.maxMemory(),
+        )
+    }
 
     @Test
     fun testAppCacheDirectory() {