+++ sketch-core/src/iosTest/kotlin/com/github/panpf/sketch/core/ios/test/util/SystemCallbacksIosTest.kt
@@ -1,0 +1,16 @@
+package com.github.panpf.sketch.core.ios.test.util
+
+import kotlin.test.Test
+
+class SystemCallbacksIosTest {
+
+    @Test
+    fun testSystemCallbacks() {
+        // TODO test
+    }
+
+    @Test
+    fun testIosSystemCallbacks() {
+        // TODO test
+    }
+}