+++ sketch-core/src/jsCommonTest/kotlin/com/github/panpf/sketch/core/jscommon/test/util/SystemCallbacksJsCommonTest.kt
@@ -1,0 +1,16 @@
+package com.github.panpf.sketch.core.jscommon.test.util
+
+import kotlin.test.Test
+
+class SystemCallbacksJsCommonTest {
+
+    @Test
+    fun testSystemCallbacks() {
+        // TODO test
+    }
+
+    @Test
+    fun testJsSystemCallbacks() {
+        // TODO test
+    }
+}