+++ sketch-core/src/nonAndroidTest/kotlin/com/github/panpf/sketch/core/nonandroid/test/util/PlatformContextsNonAndroidTest.kt
@@ -8,9 +8,4 @@ class PlatformContextsNonAndroidTest {
     fun testApplication() {
         // TODO test
     }
-
-    @Test
-    fun testTotalAvailableMemoryBytes() {
-        // TODO test
-    }
 }