+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Translations: [简体中文](CHANGELOG_zh.md)
 * change: Decoder's decode() method removes the suspend modifier and changes the return type from
   Result<DecodeResult> to DecodeResult
 * change: Transformation's transform() method removes the suspend modifier
+* improve: Desktop platforms now use 20% of Runtime.maxMemory() as the memory cache maximum
 
 # 4.0.0-alpha08